requirements.txt
appdirs==1.4.3
djangorestframework==3.6.2
packaging==16.8
pyparsing==2.2.0
six==1.10.0
